+++ spec/examples/allow_nil_on_input_with_type_and_default_nil.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AllowNilOnInputWithTypeAndDefaultNil < Actor
+  input :name, type: String, default: nil
+
+  def call; end
+end