+++ CHANGELOG.md
@@ -6,7 +6,7 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## Unreleased
 
 Breaking changes:
-- Disallow nil when a type is set by default.
+- Disallow nil when a type is set by default and the default is not nil.
 
 Added:
 - Move all code inside `ServiceActor`, only exposing a base `Actor`, enabling